+++ packages/services/src/services/Translation/Translation.js
@@ -25,7 +25,7 @@ const _host =
  * @private
  */
 const _c4dEndpointDefault =
-  '/common/carbon-for-ibm-dotcom/translations/masthead-footer';
+  '/common/carbon-for-ibm-dotcom/translations/masthead-footer/v2.1';
 
 /**
  * Translation API endpoint