+++ packages/web-components/src/components/filter-panel/filter-panel-composite.ts
@@ -8,28 +8,29 @@
  */
 
 import { html, LitElement, TemplateResult } from 'lit';
-import { property, state } from 'lit/decorators.js';
+import { property, state, query } from 'lit/decorators.js';
 import Filter from '../../internal/vendor/@carbon/web-components/icons/filter/16.js';
 import HostListenerMixin from '../../internal/vendor/@carbon/web-components/globals/mixins/host-listener.js';
-import { baseFontSize, breakpoints } from '@carbon/layout';
 import './filter-group';
 import './filter-panel';
 import './filter-panel-modal';
-import { unsafeHTML } from 'lit/directives/unsafe-html.js';
 import HostListener from '../../internal/vendor/@carbon/web-components/globals/decorators/host-listener.js';
 import settings from '../../internal/vendor/@carbon/ibmdotcom-utilities/utilities/settings/settings';
 import StableSelectorMixin from '../../globals/mixins/stable-selector';
 import styles from './filter-panel.scss';
 import '../../internal/vendor/@carbon/web-components/components/checkbox/checkbox.js';
+import '../../internal/vendor/@carbon/web-components/components/button/button.js';
 import C4DFilterGroupItem from './filter-group-item';
 import C4DFilterPanelCheckbox from './filter-panel-checkbox';
 import C4DFilterPanelInputSelect from './filter-panel-input-select';
 import C4DFilterPanelInputSelectItem from './filter-panel-input-select-item';
 import { carbonElement as customElement } from '../../internal/vendor/@carbon/web-components/globals/decorators/carbon-element.js';
+import MediaQueryMixin, {
+  MQBreakpoints,
+  MQDirs,
+} from '../../component-mixins/media-query/media-query';
 
 const { prefix, stablePrefix: c4dPrefix } = settings;
-
-const breakpoint = parseFloat(breakpoints.lg.width) * baseFontSize;
 
 /**
  * Filter panel composite
@@ -37,8 +38,9 @@ const breakpoint = parseFloat(breakpoints.lg.width) * baseFontSize;
  * @element c4d-filter-panel-composite
  */
 @customElement(`${c4dPrefix}-filter-panel-composite`)
-class C4DFilterPanelComposite extends HostListenerMixin(
-  StableSelectorMixin(LitElement)
+class C4DFilterPanelComposite extends MediaQueryMixin(
+  HostListenerMixin(StableSelectorMixin(LitElement)),
+  { [MQBreakpoints.LG]: MQDirs.MAX }
 ) {
   /**
    * Host listener for handling the state change when a input select item is selected.
@@ -64,7 +66,7 @@ class C4DFilterPanelComposite extends HostListenerMixin(
     }
 
     if (!this._selectedValues.includes(value)) {
-      this._selectedValues.push(value);
+      this._selectedValues = [...this._selectedValues, value];
     }
 
     if (lastValue && this._selectedValues.includes(lastValue)) {
@@ -72,7 +74,6 @@ class C4DFilterPanelComposite extends HostListenerMixin(
         (e) => e !== lastValue
       );
     }
-    this.renderStatus();
   };
 
   /**
@@ -107,14 +108,12 @@ class C4DFilterPanelComposite extends HostListenerMixin(
         }
       });
 
-    // toggle value in list
+    // Toggle value in list.
     if (!this._selectedValues.includes(value)) {
-      this._selectedValues.push(value);
+      this._selectedValues = [...this._selectedValues, value];
     } else {
       this._selectedValues = this._selectedValues.filter((e) => e !== value);
     }
-
-    this.renderStatus();
   };
 
   /**
@@ -132,7 +131,10 @@ class C4DFilterPanelComposite extends HostListenerMixin(
     if (match !== -1) {
       this._filterGroupsAllRevealed[match].value = event.detail.value;
     } else {
-      this._filterGroupsAllRevealed.push(event.detail);
+      this._filterGroupsAllRevealed = [
+        ...this._filterGroupsAllRevealed,
+        event.detail,
+      ];
     }
   };
 
@@ -160,7 +162,7 @@ class C4DFilterPanelComposite extends HostListenerMixin(
     // toggle checkbox in filter panel modal
     this.querySelectorAll(`${c4dPrefix}-filter-panel-input-select`).forEach(
       (e) => {
-        // capture the element counterpart in Filter Panel Modal
+        // Capture the element counterpart in Filter Panel Modal.
         if (e.getAttribute('header-value') === headerValue) {
           const currentGroup = e.closest(`${c4dPrefix}-filter-group-item`);
           currentGroup?.setAttribute('open', '');
@@ -183,20 +185,19 @@ class C4DFilterPanelComposite extends HostListenerMixin(
       }
     );
 
-    // toggle value in list
+    // Toggle value in list.
     if (!this._selectedValues.includes(headerValue)) {
-      this._selectedValues.push(headerValue);
+      this._selectedValues = [...this._selectedValues, headerValue];
     } else {
       this._selectedValues = this._selectedValues.filter(
         (e) => e !== headerValue
       );
     }
-    this.renderStatus();
   };
 
   @HostListener('document:eventHeadingChange')
   protected _handleHeadingChange = () => {
-    this.renderStatus();
+    this._setHeadingFromSlot(this._headingSlot);
   };
 
   /**
@@ -214,7 +215,7 @@ class C4DFilterPanelComposite extends HostListenerMixin(
   protected _handleClearSelection = () => {
     this._selectedValues = [];
 
-    // handles clear when clearing from the static filter panel modal
+    // Handles clear when clearing from the static filter panel modal.
     this._contents.forEach((group) => {
       group
         .querySelectorAll(`${c4dPrefix}-filter-panel-checkbox`)
@@ -253,9 +254,10 @@ class C4DFilterPanelComposite extends HostListenerMixin(
         e.removeAttribute('selected');
         e.removeAttribute('is-open');
       });
-
-    this.renderStatus();
   };
+
+  @query('slot[name=heading]')
+  _headingSlot;
 
   /**
    * `true` to open the locale modal.
@@ -279,10 +281,10 @@ class C4DFilterPanelComposite extends HostListenerMixin(
   _contents: any[] = [];
 
   /**
-   * sets the array for the filter button title
+   * Sets the heading text, to be composed with the current count.
    */
   @property()
-  _title: any[] = [];
+  _heading: string = '';
 
   /**
    * sets the selected values into an array
@@ -291,19 +293,13 @@ class C4DFilterPanelComposite extends HostListenerMixin(
   _selectedValues: string[] = [];
 
   /**
-   * sets the filter button title
-   */
-  @property()
-  _filterButtonTitle: string = '';
-
-  /**
    * stores which filter groups have revealed filters
    */
   @property()
   _filterGroupsAllRevealed: { id: string; value: boolean }[] = [];
 
-  @property()
-  _isMobile: boolean = window.innerWidth < breakpoint;
+  @state()
+  _isMobile = this.carbonBreakpoints.lg.matches;
 
   /**
    * An element to set focus to on render.
@@ -311,11 +307,9 @@ class C4DFilterPanelComposite extends HostListenerMixin(
   @state()
   _focusElement: string | null = null;
 
-  @HostListener('window:resize')
-  // @ts-ignore: The decorator refers to this method but TS thinks this method is not referred to
-  private _handleWindowResize = (): void => {
-    this._isMobile = window.innerWidth < breakpoint;
-  };
+  protected mediaQueryCallbackMaxLG() {
+    this._isMobile = this.carbonBreakpoints.lg.matches;
+  }
 
   protected async _querySelectorMobile(id: string): Promise<Element | null> {
     return this.querySelector(id);
@@ -347,16 +341,16 @@ class C4DFilterPanelComposite extends HostListenerMixin(
    * @param event The event.
    */
   protected _handleSlotChange({ target }: Event) {
-    const contents = (this._contents = (target as HTMLSlotElement)
+    this._contents = (target as HTMLSlotElement)
       .assignedNodes()
       .filter(
         (node) => node.nodeType !== Node.TEXT_NODE || node!.textContent!.trim()
-      ));
+      );
     // Calculate initial this._selectedValues. Look at the first node, which is
     // expected to be <c4d-filter-group>.
-    if (contents[0] instanceof Element) {
+    if (this._contents[0] instanceof Element) {
       const items = Array.from(
-        contents[0].querySelectorAll(
+        this._contents[0].querySelectorAll(
           `${c4dPrefix}-filter-panel-checkbox[checked],
             ${c4dPrefix}-filter-panel-input-select[selected],
             ${c4dPrefix}-filter-panel-input-select-item[selected]`
@@ -364,46 +358,72 @@ class C4DFilterPanelComposite extends HostListenerMixin(
       );
       this._selectedValues = items
         .map((item) => {
-          return item.getAttribute('value') ?? '';
+          return (
+            item.getAttribute('value') ??
+            item.getAttribute('header-value') ??
+            ''
+          );
         })
         .filter((item) => !!item);
     }
   }
 
-  protected renderStatus() {
-    this._filterButtonTitle = `
-      ${this._title[0].innerText}${
-      this._selectedValues.length > 0 ? ` (${this._selectedValues.length})` : ''
-    }
-    `;
+  protected _getComposedHeadingFilterCount() {
+    const filterCount =
+      this._selectedValues.length > 0
+        ? ` (${this._selectedValues.length})`
+        : '';
+    return `${this._heading}${filterCount}`;
   }
 
   /**
-   * Handles `slotchange` event.
+   * Handles `slotchange` event for the heading slot.
    *
    * @param event The event.
    */
-  protected _handleTitleSlotChange({ target }: Event) {
-    this._title = (target as HTMLSlotElement)
+  protected _handleHeadingSlotChange({ target }: Event) {
+    this._setHeadingFromSlot(target as HTMLSlotElement);
+  }
+
+  protected _setHeadingFromSlot(slot: HTMLSlotElement) {
+    // Clean slate.
+    this._heading = '';
+
+    // Work through candidate headings, ultimately we're only interested in the
+    // first non-empty node.
+    const candidateHeadings = slot
       .assignedNodes()
       .filter(
         (node) => node.nodeType !== Node.TEXT_NODE || node!.textContent!.trim()
       );
-    this._filterButtonTitle = this._title[0].innerText;
+
+    // If we found something, lets retain it's textContent as the heading.
+    if (candidateHeadings.length > 0) {
+      this._heading = candidateHeadings[0].textContent
+        ? candidateHeadings[0].textContent.trim()
+        : '';
+    }
   }
 
   /**
    * Renders original content into the modal and listens for changes to this
-   * content to then be stored in `this._title` and `this._content`.
+   * content to then be stored in `this._heading` and `this._contents`.
    */
   protected _renderModal = (): TemplateResult => html`
     <c4d-filter-panel-modal
       ?open=${this.openFilterModal}
-      heading="${this._filterButtonTitle}"
+      heading="${this._getComposedHeadingFilterCount()}"
       ?has-selections="${this._selectedValues.length}">
-      <slot name="heading" @slotchange="${this._handleTitleSlotChange}"></slot>
       <slot @slotchange="${this._handleSlotChange}"></slot>
     </c4d-filter-panel-modal>
+  `;
+
+  protected _renderMobile = (): TemplateResult => html`
+    <cds-button kind="tertiary" @click=${this._openModal}>
+      ${this._getComposedHeadingFilterCount()} ${Filter({ slot: 'icon' })}
+    </cds-button>
+
+    ${this._renderModal()}
   `;
 
   /**
@@ -411,25 +431,23 @@ class C4DFilterPanelComposite extends HostListenerMixin(
    */
   protected _renderDesktop = (): TemplateResult => html`
     <c4d-filter-panel
-      heading="${this._filterButtonTitle}"
+      heading="${this._getComposedHeadingFilterCount()}"
       ?has-selections="${this._selectedValues.length}">
-      ${this._title.map((e) => {
-        return html` ${unsafeHTML((e as HTMLElement).outerHTML)} `;
-      })}
-      ${this._contents.map((e) => {
-        return html` ${unsafeHTML((e as HTMLElement).outerHTML)} `;
-      })}
+      <slot @slotchange="${this._handleSlotChange}"></slot>
     </c4d-filter-panel>
   `;
 
   render() {
+    // Note that the <slot name="heading"> contents, intended to be
+    // <dds-filter-panel-heading> are never shown as is. The text contents
+    // are composed, using this._getComposedHeadingFilterCount(), together with
+    // the current filter count, and passed as an attribute to
+    // <dds-filter-panel-modal> and <dds-filter-panel>.
     return html`
-      <button class="cds--filter-button" @click=${this._openModal}>
-        <div class="${prefix}--filter__modal__button">
-          ${this._filterButtonTitle} ${Filter()}
-        </div>
-      </button>
-      ${this._renderModal()} ${this._renderDesktop()}
+      <slot
+        name="heading"
+        @slotchange="${this._handleHeadingSlotChange}"></slot>
+      ${this._isMobile ? this._renderMobile() : this._renderDesktop()}
     `;
   }