+++ packages/web-components/tests/e2e-storybook/cypress/integration/cta/card-link.e2e.js
@@ -1,5 +1,5 @@
 /**
- * Copyright IBM Corp. 2021, 2022
+ * Copyright IBM Corp. 2021, 2023
  *
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
@@ -93,7 +93,7 @@ const _tests = {
 
       if (sheets) {
         const hover = getCssPropertyForRule(
-          '.bx--card:hover, :host(cds-card:hover), :host(cds-link-list-item-card:hover), :host(cds-card-group-item:hover), :host(cds-card-group-item) .bx--card:hover, :host(cds-card-cta:hover), :host(cds-link-list-item-card-cta:hover), :host(cds-card-in-card:hover), :host(cds-content-group-cards-item:hover), :host(cds-content-group-cards-item) .bx--card:hover',
+          '.bx--card:hover, :host(cds-card:hover), :host(cds-card-group-item:hover), :host(cds-card-group-item) .bx--card:hover, :host(cds-card-cta:hover), :host(cds-card-in-card:hover), :host(cds-content-group-cards-item:hover), :host(cds-content-group-cards-item) .bx--card:hover',
           'background-color',
           sheets
         );